client: stop some taskrunner hooks when task exits
Non-sidecar prestart and poststart tasks run to completion and exit, but
their Stop hook are not called until the job is stopped. This leaves
some background operations running even when the task itself is not
running anymore.

Implementing the Exited hook stops these background processes on task
exit. If the task is restarted their Prestart and Poststart hooks will
run again and resume the background operations.