qemu: add binary_path plugin config parameter
Fixes #4684

This contains 3 changes to the qemu driver:

1. Adds a `binary_path` plugin configuration parameter for specifying
   alternative binary locations for qemu.
2. Defaults `binary_path = "qemu-system-ARCH"
3. Replaces ARCH in `binary_path` with the Nomad binary's GOARCH.

This should allow qemu to Just Work on non-x86_64 architectures like
Apple's M1/M2 or MIPS if... I honestly don't know why you'd be using
this on MIPS but it's a target both Go and QEMU support!

The `ARCH` replacement is ugly, and I had planned on using standard
`${ARCH}` replacement. However that causes it to be handled by our HCL
parser which fails with:

```
==> Error starting agent: failed to create plugin loader: failed to initialize plugin loader: parsing plugin configurations failed: 1 error occurred:
        * plugin "qemu" (driver): failed to parse config:  Variables not allowed: Variables may not be used here.
```

This is returned from `hclutils.ParseHclInterface`, and I'm not sure
it's worth fighting that battle.