authenticate: remove fast path for NodeID
Previously-connected clients will have a NodeID set on the context, and because
this is a large portion of the RPCs sent we fast-pathed it at the top of the
`Authenticate` method. But the context is shared for all yamux streams over the
same yamux session (and TCP connection). This lets an authenticated HTTP request
to a client use the NodeID for authentication, which is a privilege
escalation. Remove the fast path and annotate it so that we don't break it
again.