authenticate: add context to decisions around AuthenticatedIdentity
The `Authenticate` method taken on its own looks like it wants to return an
`acl.ACL` that folds over all the various identity types (creating an ephemeral
ACL on the fly if neccessary). But keeping these fields idependent allows RPC
handlers to differentiate between internal and external origins so we most
likely want to avoid this. Leave some docstrings as a warning as to why this is
built the way it is.