Allow per_alloc to be used with host volumes
Disallowing per_alloc for host volumes in some cases makes life of a nomad user much harder.
When we rely on the NOMAD_ALLOC_INDEX for any configuration that needs to be re-used across
restarts we need to make sure allocation placement is consistent. With CSI volumes we can
use the `per_alloc` feature but for some reason this is explicitly disabled for host volumes.

I couldn't find a reason why this done since as far as I could see the per_alloc is only used
for the scheduler to determine placement.